Added generics method to save enumerable
Internally we use DataSet and these methods are effectively just helpers that permit the easy creation of a spreadsheet from an IEnumerable.

1. public void WriteOdsFile<T>(Dictionary<string, IEnumerable<T>> sheets, string outputFilePath)
2. public void WriteOdsFile<T>(IEnumerable<T> sheet, string name, string outputFilePath)

The first method will create a number of sheets using the dictionary key as the sheet name.

The second method will create a single sheet using the name passed in.